Stub creating a view
It's handy to use the console to make sure that
the basics work.

The next question in my mind is what will make the
view decide to display a the tree div that starts
off hidden.

You might think "Oh, a click on the button" but
that's the wrong way of thinking about it.  When
the **model** that the view interfaces to updates
reaches some condition i.e. that a tree has been
planted, the tree should display.

The click of a button should *actuate* the
planting of the tree in the model, but the View
and only the view bears the responsibility for
doing the jQuery or CSS work required to display.