Create OrangesApp.Binder class
Rougly tantamount to a Ruby `Module` this is a
class which cannot be instantiated.  It's really,
as such, a namespace.

I want this class to bear all the responsibility
for setting up the bindings to click events.
`bind()` will be called but it, in turn, will call
descriptive events that do the final jQuery-level
binding.