Implement OrangesApp.Controller
It's clear that I don't want the `Binder` to bear
responsibility for handling the events.  Its
purpose is to bind things.

I want a **controller** to have the responsibilty
for dispatching events (where necessary) or
updating itself (and then calling `update()` on
the View).

At this point I can click the button and *viola* I
see a `console.log` message.

Next move, take logging out and replace it with a
change that updates the View.